Add Put Endpoint

+++ training-plan/src/main/java/learning/trainingPlan/repository/TrainingPlanRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TrainingPlanRepository extends JpaRepository<TrainingPlanEntity, Integer> {
+public interface TrainingPlanRepository extends JpaRepository<TrainingPlanEntity, Long> {
     List<TrainingPlanEntity> findFirst10ByOrderById();
 }
